requirements.txt
Django==1.10.5
django-cors-headers==2.1.0
django-filter==0.15.0
djangorestframework==3.4.7
